refactor(ui): extract combobox state management logic  into custom hook
