+++ packages/ui-components/src/components/ComboBox/hooks/index.ts
@@ -7,3 +7,5 @@ export {
 export { default as useComboBoxFloating, type ComboBoxFloatingReturn } from "./useComboBoxFloating"
 
 export { default as useComboBoxOptionFiltering } from "./useComboBoxOptionFiltering"
+
+export { default as useComboBoxState } from "./useComboBoxState"
